Move BlockNumberOrHash to eth_api module
