+++ src/rpc_client/eth_ops.rs
@@ -3,8 +3,6 @@
 
 use crate::rpc_api::eth_api::*;
 
-use crate::eth::BlockNumberOrHash;
-
 use super::{ApiInfo, RpcRequest};
 
 impl ApiInfo {
