+++ src/tool/subcommands/api_cmd.rs
@@ -13,10 +13,10 @@ use crate::blocks::Tipset;
 use crate::blocks::TipsetKeys;
 use crate::cid_collections::CidHashSet;
 use crate::db::car::ManyCar;
-use crate::eth::{BlockNumberOrHash, Predefined};
 use crate::lotus_json::HasLotusJson;
 use crate::message::Message as _;
 use crate::rpc_api::eth_api::Address as EthAddress;
+use crate::rpc_api::eth_api::*;
 use crate::rpc_client::{ApiInfo, JsonRpcError, RpcRequest};
 use crate::shim::address::{Address, Protocol};
 use crate::shim::crypto::Signature;
